refactor(chat-ai): consolidate clipboard and URL utilities, enhance message bubble interactions
