+++ apps/DocFlow/src/app/chat-ai/_components/ChatAIPanels.tsx
@@ -52,7 +52,7 @@ import {
   compactMarkdownComponents,
 } from '@/components/business/ai/markdown-components';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import { cn } from '@/utils';
+import { cn, copyImageAsBlob, copyToClipboard, isSafeUrl, safeOpenUrl } from '@/utils';
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -61,38 +61,6 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
-
-/**
- * 将图片 URL 转为 PNG Blob 并写入剪贴板
- */
-async function copyImageAsBlob(imageUrl: string): Promise<void> {
-  const img = new Image();
-  img.crossOrigin = 'anonymous';
-
-  await new Promise<void>((resolve, reject) => {
-    img.onload = () => resolve();
-    img.onerror = () => reject(new Error('图片加载失败'));
-    img.src = imageUrl;
-  });
-
-  const canvas = document.createElement('canvas');
-  canvas.width = img.naturalWidth;
-  canvas.height = img.naturalHeight;
-
-  const ctx = canvas.getContext('2d');
-
-  if (!ctx) throw new Error('Canvas 不可用');
-  ctx.drawImage(img, 0, 0);
-
-  const pngBlob = await new Promise<Blob>((resolve, reject) => {
-    canvas.toBlob((blob) => {
-      if (blob) resolve(blob);
-      else reject(new Error('图片转换失败'));
-    }, 'image/png');
-  });
-
-  await navigator.clipboard.write([new ClipboardItem({ 'image/png': pngBlob })]);
-}
 
 interface ChatPanelProps {
   /** 模型配置信息 */
@@ -132,19 +100,6 @@ function getModelDisplayName(
   const option = models.find((opt) => opt.value === modelValue);
 
   return option?.label || modelValue;
-}
-
-/**
- * 复制文本到剪贴板
- */
-async function copyToClipboard(text: string): Promise<boolean> {
-  try {
-    await navigator.clipboard.writeText(text);
-
-    return true;
-  } catch {
-    return false;
-  }
 }
 
 /**
@@ -247,7 +202,7 @@ const MessageBubble = React.memo(
                       src={message.imageUrl}
                       alt={message.content || '生成的图片'}
                       className="w-full h-auto max-h-[400px] object-contain cursor-pointer transition-transform hover:scale-[1.02]"
-                      onClick={() => window.open(message.imageUrl, '_blank')}
+                      onClick={() => safeOpenUrl(message.imageUrl)}
                       crossOrigin="anonymous"
                     />
                     <div className="absolute top-2 right-2 flex gap-1.5 opacity-0 group-hover/img:opacity-100 transition-opacity">
@@ -268,10 +223,15 @@ const MessageBubble = React.memo(
                         <ClipboardCopy className="w-3.5 h-3.5 text-gray-600" />
                       </button>
                       <button
-                        onClick={() => {
-                          if (message.imageUrl) {
-                            navigator.clipboard.writeText(message.imageUrl);
+                        onClick={async () => {
+                          if (!message.imageUrl) return;
+
+                          const success = await copyToClipboard(message.imageUrl);
+
+                          if (success) {
                             toast.success('链接已复制');
+                          } else {
+                            toast.error('复制链接失败');
                           }
                         }}
                         className="p-1.5 bg-white/90 hover:bg-white rounded-md shadow-sm transition-colors"
@@ -280,7 +240,7 @@ const MessageBubble = React.memo(
                         <Link className="w-3.5 h-3.5 text-gray-600" />
                       </button>
                       <a
-                        href={message.imageUrl}
+                        href={isSafeUrl(message.imageUrl) ? message.imageUrl : undefined}
                         download
                         target="_blank"
                         rel="noopener noreferrer"
